+++ CHANGELOG.md
@@ -4,6 +4,15 @@
 
 ### Added
 
+- Added `Get-JiraAgileBoardIssue` for `GET /rest/agile/1.0/board/{boardId}/issue`
+- Added `Get-JiraAgileBacklogIssue` for `GET /rest/agile/1.0/board/{boardId}/backlog`
+- Added `Get-JiraAgileSprintIssue` for `GET /rest/agile/1.0/board/{boardId}/sprint/{sprintId}/issue`
+- Added `Get-JiraAgileBoardConfiguration` for `GET /rest/agile/1.0/board/{boardId}/configuration`
+- Added `Get-JiraAgileBoardEpic` for `GET /rest/agile/1.0/board/{boardId}/epic`
+- Added `Get-JiraAgileEpic` for `GET /rest/agile/1.0/epic/{epicId}`
+- Added `Get-JiraAgileEpicIssue` for epic and board-scoped epic issue retrieval endpoints
+- Added conversion helpers for paged Agile issue/epic responses and board configuration payloads
+
 ### Changed
 
 ### Fixed